--- core/src/main/scala/argonaut/Shapeless.scala
@@ -1,12 +1,0 @@
-package argonaut
-
-import argonaut.derive._
-
-object Shapeless
-  extends SingletonInstances
-  with DerivedInstances {
-
-  object Cached
-    extends SingletonInstances
-    with CachedDerivedInstances
-}
