+++ refined/src/main/scala/argonaut/ArgonautRefined.scala
@@ -1,0 +1,10 @@
+package argonaut
+
+import argonaut.derive.RefinedInstances
+
+import language.higherKinds
+
+object ArgonautRefined extends RefinedInstances
+
+@deprecated("Use argonaut.ArgonautRefined instead. Will be removed in argonaut-shapeless 1.2.")
+object Refined extends RefinedInstances
