--- refined/src/main/scala/argonaut/Refined.scala
@@ -1,7 +1,0 @@
-package argonaut
-
-import argonaut.derive.RefinedInstances
-
-import language.higherKinds
-
-object Refined extends RefinedInstances
