+++ test/src/test/scala-2.11/argonaut/Refined211EncodeTests.scala
@@ -2,7 +2,8 @@ package argonaut
 
 import eu.timepit.refined._, numeric._, collection._, boolean._
 import shapeless.nat.{ _0, _3 }
-import Argonaut._, Refined._
+import argonaut.Argonaut._
+import argonaut.ArgonautRefined._
 import utest._
 
 object Refined211EncodeTests extends TestSuite {
