+++ test/src/test/scala/argonaut/PriorityTests.scala
@@ -2,7 +2,8 @@ package argonaut
 
 import utest._
 
-import Argonaut._, Shapeless._
+import argonaut.Argonaut._
+import argonaut.ArgonautShapeless._
 
 object PriorityTests extends TestSuite {
   import PriorityTestsDefn._
