+++ test/src/test/scala/argonaut/ProductEncodeTests.scala
@@ -12,7 +12,8 @@ import derive._
 
 import org.scalacheck.Shapeless._
 
-import Argonaut._, Shapeless._
+import argonaut.Argonaut._
+import argonaut.ArgonautShapeless._
 
 
 object ProductEncodeTests extends TestSuite {
